Hunyuan3D-Part/requirements.txt
--find-links https://data.pyg.org/whl/torch-2.8.0+cu129.html
# Build Tools
ninja==1.11.1.1

# Core ML/Deep Learning
transformers==4.46.0
diffusers==0.30.0
accelerate==1.1.1
pytorch-lightning==1.9.5
huggingface-hub==0.35.1
safetensors==0.4.4

# Scientific Computing
numpy==2.1.2
scipy==1.15.3
einops==0.8.0
pandas==2.2.2

# 3D Mesh Processing
trimesh==4.4.7
pymeshlab==2022.2.post3

# Configuration Management
omegaconf==2.3.0
pyyaml==6.0.2
configargparse==1.7

# Utilities
tqdm==4.66.5

# GPU Computing (requires CUDA)
cupy-cuda12x==13.4.1

# ONNX Runtime
onnxruntime==1.16.3
torchmetrics==1.6.0

timm
numba
fpsample
scikit-learn
addict
scikit-image
easydict

# sonata
spconv-cu126
git+https://github.com/Dao-AILab/flash-attention.git
torch-scatter
torch-cluster